refactor: changed name of first coral raw dataset from 2024-->2023
This makes the range of the data clearer. I had initially suffixed the file 2024 because that was the year I received the dataset from Hunter, but the 2024 survey data was not contained. The 2024 survey data was subsequently provided but in the form of individual .xls files, which are contained in an adjacent folder named data_raw_2024_sheets.