refactor: added ex8 as a file in the cleaning script pipeline
The exploration file ex8 is where I previously took all the disparate excel files in the 2024 surveys folder and made them into a data dataset that the cleaning pipeline could utilize. Because this file represents an important step in the process, I've moved it so that it may be its own file in the cleaning pipeline.